add authRouter for user and login
